backend/requirements.txt
Django
django-rest-framework
djangorestframework
djangorestframework_simplejwt